feat: release version 1.3.8

+++ main.go
@@ -1,49 +1,59 @@
 package main
 
 import (
-	"embed"
+	"flag"
 	"fmt"
 	"io"
+	"io/ioutil"
 	"os"
 	"os/exec"
+	"os/signal"
 	"sync"
+	"syscall"
 )
 
-//go:embed assets/nancy.mp3
-var defaultMusic embed.FS
+const musicFilePath = "assets/nancy.mp3"
 
-func playMusic(wg *sync.WaitGroup, stopChan chan struct{}) {
+func playMusic(wg *sync.WaitGroup, stopChan chan struct{}, musicPath string) {
 	defer wg.Done()
 
-	tmpFile, err := os.CreateTemp("", "nancy*.mp3")
-	if err != nil {
-		fmt.Println("Failed to create temp file:", err)
-		return
-	}
-	defer os.Remove(tmpFile.Name())
+	var musicFilePath string
 
-	// Copy the embedded file to the temp file
-	file, err := defaultMusic.Open("assets/nancy.mp3")
-	if err != nil {
-		fmt.Println("Failed to open embedded music file:", err)
-		return
-	}
-	defer file.Close()
+	if musicPath != "" {
+		musicFilePath = musicPath
+	} else {
+		tmpFile, err := ioutil.TempFile("", "nancy.mp3")
+		if err != nil {
+			fmt.Println("Failed to create temp file:", err)
+			return
+		}
+		defer os.Remove(tmpFile.Name())
 
-	_, err = io.Copy(tmpFile, file)
-	if err != nil {
-		fmt.Println("Failed to copy embedded music file:", err)
-		return
-	}
+		// Open the music file from the assets directory
+		file, err := os.Open(musicFilePath)
+		if err != nil {
+			fmt.Println("Failed to open music file:", err)
+			return
+		}
+		defer file.Close()
 
-	// Close the temp file to ensure it's written
-	if err := tmpFile.Close(); err != nil {
-		fmt.Println("Failed to close temp file:", err)
-		return
+		_, err = io.Copy(tmpFile, file)
+		if err != nil {
+			fmt.Println("Failed to copy music file:", err)
+			return
+		}
+
+		// Close the temp file to ensure it's written
+		if err := tmpFile.Close(); err != nil {
+			fmt.Println("Failed to close temp file:", err)
+			return
+		}
+
+		musicFilePath = tmpFile.Name()
 	}
 
 	// Play the file using mpg123
-	cmd := exec.Command("mpg123", "-q", "--loop", "-1", tmpFile.Name())
+	cmd := exec.Command("mpg123", "-q", "--loop", "-1", musicFilePath)
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
 
@@ -52,8 +62,12 @@ func playMusic(wg *sync.WaitGroup, stopChan chan struct{}) {
 		return
 	}
 
-	<-stopChan
-	cmd.Process.Kill()
+	go func() {
+		<-stopChan
+		cmd.Process.Kill()
+	}()
+
+	cmd.Wait()
 }
 
 func runCommand(command string, args []string, wg *sync.WaitGroup, stopChan chan struct{}) {
@@ -73,19 +87,30 @@ func runCommand(command string, args []string, wg *sync.WaitGroup, stopChan chan
 }
 
 func main() {
-	if len(os.Args) < 2 {
-		fmt.Println("Usage: sahh-music-cli <command> [args...]")
+	musicPath := flag.String("music", "", "Path to custom MP3 file")
+	flag.Parse()
+
+	if len(flag.Args()) < 1 {
+		fmt.Println("Usage: sahh [--music <path_to_mp3>] <command> [args...]")
 		os.Exit(1)
 	}
 
-	command := os.Args[1]
-	commandArgs := os.Args[2:]
+	command := flag.Args()[0]
+	commandArgs := flag.Args()[1:]
 
 	var wg sync.WaitGroup
 	stopChan := make(chan struct{})
 
+	// Handle Ctrl+C
+	sigChan := make(chan os.Signal, 1)
+	signal.Notify(sigChan, syscall.SIGINT, syscall.SIGTERM)
+	go func() {
+		<-sigChan
+		close(stopChan)
+	}()
+
 	wg.Add(2)
-	go playMusic(&wg, stopChan)
+	go playMusic(&wg, stopChan, *musicPath)
 	go runCommand(command, commandArgs, &wg, stopChan)
 
 	wg.Wait()
